Removed WPMainTabLayout
